#215: finish cxx_main.cpp
